+++ android/src/main/java/com/synerise/sdk/react/RNContent.java
@@ -4,29 +4,38 @@ import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Callback;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactMethod;
+import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
+import com.facebook.react.bridge.ReadableType;
 import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
 import com.google.gson.Gson;
+import com.synerise.sdk.client.model.simpleAuth.ClientData;
 import com.synerise.sdk.content.Content;
 import com.synerise.sdk.content.model.Audience;
 import com.synerise.sdk.content.model.DocumentsApiQuery;
 import com.synerise.sdk.content.model.DocumentsApiQueryType;
 import com.synerise.sdk.content.model.ScreenViewResponse;
 import com.synerise.sdk.content.model.document.Document;
+import com.synerise.sdk.content.model.document.DocumentApiQuery;
+import com.synerise.sdk.content.model.recommendation.FiltersJoinerRule;
 import com.synerise.sdk.content.model.recommendation.RecommendationRequestBody;
 import com.synerise.sdk.content.model.recommendation.RecommendationResponse;
 import com.synerise.sdk.content.model.screenview.ScreenView;
+import com.synerise.sdk.content.model.screenview.ScreenViewApiQuery;
 import com.synerise.sdk.content.widgets.dataModel.Recommendation;
 import com.synerise.sdk.core.listeners.DataActionListener;
 import com.synerise.sdk.core.net.IDataApiCall;
 import com.synerise.sdk.error.ApiError;
+import com.synerise.sdk.promotions.model.promotion.PromotionStatus;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
@@ -191,13 +200,11 @@ public class RNContent extends RNBaseModule {
 
     @ReactMethod
     public void getRecommendationsV2(ReadableMap recommendationsOptions, Callback callback) {
-        String productId = recommendationsOptions.hasKey("productID") ? recommendationsOptions.getString("productID") : "";
         String slugName = recommendationsOptions.hasKey("slug") ? recommendationsOptions.getString("slug") : "";
 
         if (getRecommendationsApiCall != null) getRecommendationsApiCall.cancel();
 
-        RecommendationRequestBody recommendationRequestBody = new RecommendationRequestBody();
-        recommendationRequestBody.setProductId(productId);
+        RecommendationRequestBody recommendationRequestBody = readableMapToRecommendationRequestBody(recommendationsOptions);
 
         getRecommendationsApiCall = Content.getRecommendationsV2(slugName, recommendationRequestBody);
         getRecommendationsApiCall.execute(new DataActionListener<RecommendationResponse>() {
@@ -227,6 +234,37 @@ public class RNContent extends RNBaseModule {
         if (generateDocumentApiCall != null) generateDocumentApiCall.cancel();
 
         generateDocumentApiCall = Content.generateDocument(slug);
+        generateDocumentApiCall.execute(new DataActionListener<Document>() {
+
+            @Override
+            public void onDataAction(Document document) {
+                WritableMap documentMap = Arguments.createMap();
+                documentMap.putString("uuid", document.getUuid());
+                documentMap.putString("slug", document.getSlug());
+                documentMap.putString("schema", document.getSchema());
+                try {
+                    String jsonObject = gson.toJson(document);
+                    WritableMap objectMap = convertJsonToMap(new JSONObject(jsonObject));
+                    documentMap.putMap("content", objectMap);
+                    executeSuccessCallbackResponse(callback, documentMap, null);
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+            }
+        }, new DataActionListener<ApiError>() {
+            @Override
+            public void onDataAction(ApiError apiError) {
+                executeFailureCallbackResponse(callback, null, apiError);
+            }
+        });
+    }
+
+    @ReactMethod
+    public void generateDocumentWithApiQuery(ReadableMap documentApiQueryMap, Callback callback) {
+        if (generateDocumentApiCall != null) generateDocumentApiCall.cancel();
+
+        DocumentApiQuery documentApiQuery = readableMapToDocumentApiQuery(documentApiQueryMap);
+        generateDocumentApiCall = Content.generateDocument(documentApiQuery);
         generateDocumentApiCall.execute(new DataActionListener<Document>() {
 
             @Override
@@ -292,6 +330,103 @@ public class RNContent extends RNBaseModule {
                     }
                 }
         );
+    }
+
+    @ReactMethod
+    public void generateScreenViewWithApiQuery(ReadableMap screenViewApiQueryMap, Callback callback) {
+        String feedSlug = screenViewApiQueryMap.hasKey("feedSlug") ? screenViewApiQueryMap.getString("feedSlug") : "";
+        String productId = screenViewApiQueryMap.hasKey("productId") ? screenViewApiQueryMap.getString("productId") : "";
+        ScreenViewApiQuery screenViewApiQuery = new ScreenViewApiQuery(feedSlug, productId);
+        if (generateScreenViewApiCall != null) generateScreenViewApiCall.cancel();
+        generateScreenViewApiCall = Content.generateScreenView(screenViewApiQuery);
+        generateScreenViewApiCall.execute(
+                new DataActionListener<ScreenView>() {
+                    @Override
+                    public void onDataAction(ScreenView data) {
+                        WritableMap screenViewMap = Arguments.createMap();
+                        screenViewMap.putMap("audience", audienceNewScreenViewsToWritableMap(data.getAudience()));
+                        screenViewMap.putString("identifier", data.getId());
+                        screenViewMap.putString("hash", data.getHash());
+                        screenViewMap.putString("path", data.getPath());
+                        screenViewMap.putString("name", data.getName());
+                        screenViewMap.putInt("priority", data.getPriority());
+                        screenViewMap.putMap("data", screenViewDataToWritableMap(data.getData()));
+
+                        try {
+                            Date createdAtDate = new SimpleDateFormat(ISO8601_FORMAT, Locale.getDefault()).parse(data.getCreatedAt());
+                            Date updatedAtDate = new SimpleDateFormat(ISO8601_FORMAT, Locale.getDefault()).parse(data.getUpdatedAt());
+
+                            if (createdAtDate != null) {
+                                screenViewMap.putDouble("createdAt", createdAtDate.getTime());
+                            }
+                            if (updatedAtDate != null) {
+                                screenViewMap.putDouble("updatedAt", updatedAtDate.getTime());
+                            }
+                        } catch (ParseException e) {
+                            e.printStackTrace();
+                        }
+
+                        executeSuccessCallbackResponse(callback, screenViewMap, null);
+                    }
+                }, new DataActionListener<ApiError>() {
+                    @Override
+                    public void onDataAction(ApiError apiError) {
+                        executeFailureCallbackResponse(callback, null, apiError);
+                    }
+                }
+        );
+    }
+
+    private DocumentApiQuery readableMapToDocumentApiQuery(ReadableMap map) {
+        String slugName = map.hasKey("feedSlug") ? map.getString("feedSlug") : "";
+        DocumentApiQuery documentApiQuery = new DocumentApiQuery(slugName);
+        documentApiQuery.setProductId(map.hasKey("productId") ? map.getString("productId") : null);
+        documentApiQuery.setItemsIds(map.hasKey("itemsIds") ? readableArrayToListOfStrings(map.getArray("itemsIds")) : null);
+        documentApiQuery.setItemsExcluded(map.hasKey("itemsExcluded") ? readableArrayToListOfStrings(map.getArray("itemsExcluded")) : null);
+        documentApiQuery.setAdditionalFilters(map.hasKey("additionalFilters") ? map.getString("additionalFilters") : null);
+        documentApiQuery.setAdditionalElasticFilters(map.hasKey("additionalElasticFilters") ? map.getString("additionalElasticFilters") : null);
+        documentApiQuery.setDisplayAttributes(map.hasKey("displayAttribute") ? readableArrayToListOfStrings(map.getArray("displayAttribute")) : null);
+        documentApiQuery.setIncludeContextItems(map.hasKey("includeContextItems") ? map.getBoolean("includeContextItems") : null);
+        if (map.hasKey("filtersJoiner")) {
+            documentApiQuery.setFiltersJoiner(FiltersJoinerRule.valueOf(map.getString("filtersJoiner")));
+        }
+        if (map.hasKey("elasticFiltersJoiner")) {
+            documentApiQuery.setElasticFiltersJoiner(FiltersJoinerRule.valueOf(map.getString("elasticFiltersJoiner")));
+        }
+
+        return documentApiQuery;
+    }
+
+    private RecommendationRequestBody readableMapToRecommendationRequestBody(ReadableMap map) {
+        RecommendationRequestBody recommendationRequestBody = new RecommendationRequestBody();
+        recommendationRequestBody.setProductId(map.hasKey("productID") ? map.getString("productID") : null);
+        recommendationRequestBody.setItemsIds(map.hasKey("itemsIds") ? readableArrayToListOfStrings(map.getArray("itemsIds")) : null);
+        recommendationRequestBody.setItemsExcluded(map.hasKey("itemsExcluded") ? readableArrayToListOfStrings(map.getArray("itemsExcluded")) : null);
+        recommendationRequestBody.setAdditionalFilters(map.hasKey("additionalFilters") ? map.getString("additionalFilters") : null);
+        recommendationRequestBody.setAdditionalElasticFilters(map.hasKey("additionalElasticFilters") ? map.getString("additionalElasticFilters") : null);
+        recommendationRequestBody.setDisplayAttributes(map.hasKey("displayAttribute") ? readableArrayToListOfStrings(map.getArray("displayAttribute")) : null);
+        recommendationRequestBody.setIncludeContextItems(map.hasKey("includeContextItems") ? map.getBoolean("includeContextItems") : null);
+        if (map.hasKey("filtersJoiner")) {
+            recommendationRequestBody.setFiltersJoiner(FiltersJoinerRule.valueOf(map.getString("filtersJoiner")));
+        }
+        if (map.hasKey("elasticFiltersJoiner")) {
+            recommendationRequestBody.setElasticFiltersJoiner(FiltersJoinerRule.valueOf(map.getString("elasticFiltersJoiner")));
+        }
+
+        return recommendationRequestBody;
+    }
+
+    private static ArrayList<String> readableArrayToListOfStrings(ReadableArray readableArray) {
+        ArrayList<String> list = new ArrayList<>();
+
+        for (int i = 0; i < readableArray.size(); i++) {
+            ReadableType type = readableArray.getType(i);
+            if (type == ReadableType.String) {
+                list.add(readableArray.getString(i));
+            }
+        }
+
+        return list;
     }
 
     private WritableArray recommendationToWritableArray(List<Recommendation> array) {