+++ lib/classes/api_queries/ScreenViewApiQuery.d.ts
@@ -1,0 +1,10 @@
+declare class ScreenViewApiQuery {
+    feedSlug: string;
+    productId?: string | null;
+    constructor(feedSlug: string, productId?: string | null);
+    toObject(): {
+        feedSlug: string;
+        productId: string | null | undefined;
+    };
+}
+export { ScreenViewApiQuery };